W-15854291: Refactored code to collapse following hierarchy: 1. AbstractFileInputStreamSupplier and AbstractConnectedFileInputStreamSupplier - FtpFileInputStreamSupplier 2. AbstractPostActionGroup -  PostActionGroup 3. AbstractFtpCopyDelegate - FtpCopyDelegate
