+++ src/main/java/org/mule/extension/ftp/internal/source/PostActionGroup.java
@@ -6,19 +6,30 @@
  */
 package org.mule.extension.ftp.internal.source;
 
+import static java.lang.String.format;
 import static org.mule.runtime.api.meta.model.display.PathModel.Location.EXTERNAL;
 import static org.mule.runtime.api.meta.model.display.PathModel.Type.DIRECTORY;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import org.mule.extension.ftp.api.ftp.FtpFileAttributes;
+import org.mule.extension.ftp.internal.config.FileConnectorConfig;
+import org.mule.extension.ftp.internal.connection.FtpFileSystem;
+import org.mule.extension.ftp.internal.exception.FileAlreadyExistsException;
 import org.mule.runtime.extension.api.annotation.param.Optional;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.param.display.Path;
+
+import java.nio.file.Paths;
 
 /**
  * Groups post processing action parameters
  *
  * @since 1.1
  */
-public class PostActionGroup extends AbstractPostActionGroup {
+public class PostActionGroup {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(PostActionGroup.class);
 
   /**
    * Whether each file should be deleted after processing or not
@@ -92,5 +103,58 @@ public class PostActionGroup extends AbstractPostActionGroup {
 
   public boolean getOverwrite() {
     return overwrite;
+  }
+
+  public void validateSelf() throws IllegalArgumentException {
+    if (isAutoDelete()) {
+      if (getMoveToDirectory() != null) {
+        throw new IllegalArgumentException(format("The autoDelete parameter was set to true, but the value '%s' was given to the "
+            + "moveToDirectory parameter. These two are contradictory.", getMoveToDirectory()));
+      } else if (getRenameTo() != null) {
+        throw new IllegalArgumentException(format("The autoDelete parameter was set to true, but the value '%s' was given to the "
+            + "renameTo parameter. These two are contradictory.", getRenameTo()));
+      }
+    }
+  }
+
+  public void apply(FtpFileSystem fileSystem, FtpFileAttributes ftpFileAttributes, FileConnectorConfig config) {
+    if (LOGGER.isTraceEnabled()) {
+      try {
+        validateSelf();
+      } catch (IllegalArgumentException e) {
+        LOGGER.trace(e.getMessage());
+      }
+    }
+
+    boolean movedOrRenamed = false;
+    try {
+      if (getMoveToDirectory() != null) {
+        fileSystem.move(config, ftpFileAttributes.getPath(), getMoveToDirectory(), getOverwrite(), true,
+                        getRenameTo());
+        movedOrRenamed = true;
+      } else if (getRenameTo() != null) {
+        fileSystem.rename(ftpFileAttributes.getPath(), getRenameTo(), getOverwrite());
+        movedOrRenamed = true;
+      }
+    } catch (FileAlreadyExistsException e) {
+      if (!isAutoDelete()) {
+        if (getMoveToDirectory() == null) {
+          LOGGER.warn(format("A file with the same name was found when trying to rename '%s' to '%s'" +
+              ". The file '%s' was not renamed and it remains on the poll directory.",
+                             ftpFileAttributes.getFileName(), getRenameTo(), ftpFileAttributes.getPath()));
+        } else {
+          String moveToFileName = getRenameTo() == null ? ftpFileAttributes.getFileName() : getRenameTo();
+          String moveToPath = Paths.get(getMoveToDirectory()).resolve(moveToFileName).toString();
+          LOGGER.warn(format("A file with the same name was found when trying to move '%s' to '%s'" +
+              ". The file '%s' was not sent to the moveTo directory and it remains on the poll directory.",
+                             ftpFileAttributes.getPath(), moveToPath, ftpFileAttributes.getPath()));
+        }
+        throw e;
+      }
+    } finally {
+      if (isAutoDelete() && !movedOrRenamed) {
+        fileSystem.delete(ftpFileAttributes.getPath());
+      }
+    }
   }
 }
